requirements.txt
python-dotenv==1.0.1
typer==0.9.0
loguru==0.7.2
requests==2.32.3
pydantic==2.8.2
scikit-learn==1.5.1
python-docx==1.1.2
beautifulsoup4==4.12.3
numpy~=2.3.3
urllib3~=2.5.0
PyPDF2==3.0.1
pdfminer.six
pdf2image
pytesseract
sentence-transformers==2.7.0
click==8.1.7
streamlit